requirements.txt
python-dotenv
langchain-openai
langchain-core
langchain-community
flask
beautifulsoup4
